+++ corporal/connector/synapse.go
@@ -5,6 +5,7 @@ import (
 	"devture-matrix-corporal/corporal/util"
 	"fmt"
 	"strings"
+	"sync"
 
 	"crypto/hmac"
 	"crypto/sha1"
@@ -12,22 +13,120 @@ import (
 	"github.com/matrix-org/gomatrix"
 )
 
+const (
+	deviceIdCorporal = "matrix-corporal"
+)
+
 // SynapseConnector is a MatrixConnector implementation for controlling a Synapse server.
 // It is based on the base ApiConnector for doing whatever's possible,
 // but also contains Synapse-specific API calls here.
 type SynapseConnector struct {
 	*ApiConnector
+
 	registrationSharedSecret string
+	corporalUserID           string
+
+	corporalUserAccessTokenContext *AccessTokenContext
+
+	corporalUserIDLock *sync.Mutex
 }
 
 func NewSynapseConnector(
 	apiConnector *ApiConnector,
 	registrationSharedSecret string,
+	corporalUserID string,
 ) *SynapseConnector {
-	return &SynapseConnector{
-		ApiConnector:             apiConnector,
+	me := &SynapseConnector{
+		ApiConnector: apiConnector,
+
 		registrationSharedSecret: registrationSharedSecret,
+		corporalUserID:           corporalUserID,
+
+		corporalUserIDLock: &sync.Mutex{},
 	}
+
+	// This is a special access token context that we only use for the matrix-corporal user.
+	// We force it to use the ApiConnector (and not this SynapseConnector),
+	// because we wish to obtain a token for that user directly and not via the custom admin APIs in `ObtainNewAccessTokenForUserId()` below.
+	me.corporalUserAccessTokenContext = NewAccessTokenContext(me.ApiConnector, deviceIdCorporal)
+
+	return me
+}
+
+func (me *SynapseConnector) getAccessTokenForCorporalUser() (string, error) {
+	me.corporalUserIDLock.Lock()
+	defer me.corporalUserIDLock.Unlock()
+
+	return me.corporalUserAccessTokenContext.GetAccessTokenForUserId(me.corporalUserID)
+}
+
+// ObtainNewAccessTokenForUserId is a reimplementation of ApiConnector.ObtainNewAccessTokenForUserId.
+//
+// ApiConnector.ObtainNewAccessTokenForUserId uses the regular `/_matrix/client/r0/login` endpoint
+// and relies on shared-secret-auth to impersonate a user.
+//
+// This implementation here relies on an admin's access token and on the `POST /_synapse/admin/v1/users/<user_id>/login` API
+// (see https://github.com/matrix-org/synapse/pull/8617), to obtain a non-device-creating token for any user.
+//
+// Not creating devices leads to better performance and UX (no need to notify others via federation; the user's device list does not get poluted).
+// This is Synapse-specific though.
+func (me *SynapseConnector) ObtainNewAccessTokenForUserId(userId, deviceId string) (string, error) {
+	if userId == me.corporalUserID {
+		// Someone explicitly requested a token for the matrix-corporal user.
+		// If we try to proceed below (using the Admin user login API to log in as matrix-corporal),
+		// we'll hit an error: "Cannot use admin API to login as self".
+		//
+		// Requests to log in as the matrix-corporal user should be handled separately.
+		//
+		// We may wish to use `me.getAccessTokenForCorporalUser()` and just return that, but that's
+		// also not good enough, because we use this token for our own internal purposes and we want it
+		// to remain valid until we dispose of it ourselves.
+		// Giving it out to consumers means that they may destroy it (and our reconciliation code will certainly do that during cleanup!).
+		// Having this token end up destroyed will break all future invocations of this function.
+		//
+		// So.. we don't reuse an existing token, but always obtain a fresh one.
+		return me.ApiConnector.ObtainNewAccessTokenForUserId(userId, deviceId)
+	}
+
+	corporalUserAccessToken, err := me.getAccessTokenForCorporalUser()
+	if err != nil {
+		return "", fmt.Errorf(
+			"Could not obtain access token for `%s`, necessary for obtaining a token for `%s`: %s",
+			me.corporalUserID,
+			userId,
+			err,
+		)
+	}
+
+	client, err := me.createMatrixClientForUserIdAndToken(me.corporalUserID, corporalUserAccessToken)
+	if err != nil {
+		return "", err
+	}
+
+	loginEndpoint := fmt.Sprintf(
+		"/_synapse/admin/v1/users/%s/login",
+		userId,
+	)
+
+	// TODO - using `valid_until_ms` may be a great idea.
+	// However, we shouldn't do it for every token obtain, but rather only if the caller asks us to.
+	//
+	// Reconciliation should use some generous validity (but expire nevertheless),
+	// while our own HTTP APIs for obtaining/releasing tokens may do something else.
+	requestPayload := map[string]string{}
+
+	url := client.BuildURLWithQuery([]string{loginEndpoint}, map[string]string{})
+	// The URL-building function above forces us under the `/_matrix/client/r0/` prefix.
+	// We'd like to work at the top-level though, hence this hack.
+	url = strings.Replace(url, "/_matrix/client/r0/", "/", 1)
+
+	var response matrix.ApiAdminResponseUserLogin
+	err = client.MakeRequest("POST", url, requestPayload, &response)
+	if err != nil {
+		return "", err
+	}
+
+	return response.AccessToken, nil
 }
 
 func (me *SynapseConnector) DetermineCurrentState(
@@ -173,4 +272,8 @@ func (me *SynapseConnector) EnsureUserAccountExists(userId, password string) err
 	clientForUser.Logout()
 
 	return nil
+}
+
+func (me *SynapseConnector) Release() {
+	me.corporalUserAccessTokenContext.Release()
 }